Create abstract.java
abtract class cannot be called directly. it can call from an inherited child class. 

getClass() is the method of Object class. This method returns the runtime class of this object. The class object which is returned is the object that is locked by static synchronized method of the represented class.